Make capitalisation of Findbugs consistent
